+++ src/main/java/uk/gov/hmcts/reform/prl/services/PaymentRequestService.java
@@ -1,42 +1,43 @@
 package uk.gov.hmcts.reform.prl.services;
 
+import com.fasterxml.jackson.databind.ObjectMapper;
 import lombok.RequiredArgsConstructor;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.stereotype.Component;
 import org.springframework.stereotype.Service;
 import uk.gov.hmcts.reform.authorisation.generators.AuthTokenGenerator;
-import uk.gov.hmcts.reform.ccd.client.model.CallbackRequest;
-import uk.gov.hmcts.reform.prl.controllers.AbstractCallbackController;
 import uk.gov.hmcts.reform.prl.enums.OrchestrationConstants;
 import uk.gov.hmcts.reform.prl.models.FeeResponse;
 import uk.gov.hmcts.reform.prl.models.FeeType;
+import uk.gov.hmcts.reform.prl.models.dto.ccd.CallbackRequest;
 import uk.gov.hmcts.reform.prl.models.dto.ccd.CaseData;
 import uk.gov.hmcts.reform.prl.models.dto.payment.CasePaymentRequestDto;
 import uk.gov.hmcts.reform.prl.models.dto.payment.FeeDto;
 import uk.gov.hmcts.reform.prl.models.dto.payment.PaymentServiceRequest;
 import uk.gov.hmcts.reform.prl.models.dto.payment.PaymentServiceResponse;
 
-import java.math.BigDecimal;
-
 @Service
 @Component
 @RequiredArgsConstructor(onConstructor = @__(@Autowired))
-public class PaymentRequestService extends AbstractCallbackController {
+public class PaymentRequestService {
 
-    @Autowired
+
     private final PaymentApi paymentApi;
-
-    @Autowired
     private final AuthTokenGenerator authTokenGenerator;
     private final FeeService feeService;
+    private final ObjectMapper objectMapper;
 
     @Value("${payments.api.callback-url}")
     String callBackURL;
 
 
     public PaymentServiceResponse createServiceRequest(CallbackRequest callbackRequest, String authorisation) throws Exception {
-        CaseData caseData = getCaseData(callbackRequest.getCaseDetails());
+        CaseData caseData = objectMapper.convertValue(
+            CaseData.builder().applicantCaseName(callbackRequest.getCaseDetails().getCaseData().getApplicantCaseName())
+            .id(Long.valueOf(callbackRequest.getCaseDetails().getCaseId())).build(),
+            CaseData.class
+        );
         FeeResponse feeResponse= feeService.fetchFeeDetails(FeeType.C100_SUBMISSION_FEE);
         PaymentServiceResponse paymentServiceResponse = paymentApi
             .createPaymentServiceRequest(authorisation, authTokenGenerator.generate(), PaymentServiceRequest.builder()
