+++ src/main/java/uk/gov/hmcts/reform/prl/services/RequestUpdateCallbackService.java
@@ -37,12 +37,18 @@ public class RequestUpdateCallbackService {
 
     public void processCallback(ServiceRequestUpdateDto serviceRequestUpdateDto) throws Exception {
 
-        log.info("Processing the callback for the caseId {} with status {}",serviceRequestUpdateDto.getCcdCaseNumber(),
-                 serviceRequestUpdateDto.getServiceRequestStatus());
+        log.info("Processing the callback for the caseId {} with status {}", serviceRequestUpdateDto.getCcdCaseNumber(),
+                 serviceRequestUpdateDto.getServiceRequestStatus()
+        );
         String userToken = systemUserService.getSysUserToken();
         String systemUpdateUserId = systemUserService.getUserId(userToken);
-        CaseDetails caseDetails = coreCaseDataApi.getCase(userToken, authTokenGenerator.generate(), serviceRequestUpdateDto.getCcdCaseNumber());
-        if (!Objects.isNull(caseDetails.getId()) && serviceRequestUpdateDto.getServiceRequestStatus().equalsIgnoreCase(PAID)) {
+        CaseDetails caseDetails = coreCaseDataApi.getCase(
+            userToken,
+            authTokenGenerator.generate(),
+            serviceRequestUpdateDto.getCcdCaseNumber()
+        );
+        if (!Objects.isNull(caseDetails.getId()) && serviceRequestUpdateDto.getServiceRequestStatus().equalsIgnoreCase(
+            PAID)) {
             CaseData caseData = setCaseData(serviceRequestUpdateDto);
             StartEventResponse startEventResponse = coreCaseDataApi.startEventForCaseWorker(
                 userToken,
@@ -72,8 +78,7 @@ public class RequestUpdateCallbackService {
                 true,
                 caseDataContent
             );
-        }
-        else{
+        } else {
             log.error("Case id {} not present or status is not paid", serviceRequestUpdateDto.getCcdCaseNumber());
             throw new Exception("Case not present or status is not paid");
         }
@@ -81,22 +86,19 @@ public class RequestUpdateCallbackService {
 
     private CaseData setCaseData(ServiceRequestUpdateDto serviceRequestUpdateDto) {
         return objectMapper.convertValue(CaseData.builder()
-                                             .id(Long.valueOf(serviceRequestUpdateDto.getCcdCaseNumber()))
-                                             .paymentCallbackServiceRequestUpdate(
-                                                 CcdPaymentServiceRequestUpdate.builder()
-                                                     .serviceRequestReference(
-                                                         serviceRequestUpdateDto.getServiceRequestReference())
-                                                     .ccdCaseNumber(serviceRequestUpdateDto.getCcdCaseNumber())
-                                                     .serviceRequestAmount(serviceRequestUpdateDto.getServiceRequestAmount())
-                                                     .serviceRequestStatus(serviceRequestUpdateDto.getServiceRequestStatus())
-                                                     .callBackUpdateTimestamp(LocalDateTime.now())
-                                                     .payment(CcdPayment.builder().paymentAmount(
-                                                         serviceRequestUpdateDto.getPayment().getPaymentAmount())
-                                                                  .paymentReference(serviceRequestUpdateDto.getPayment().getPaymentReference())
-                                                                  .paymentMethod(serviceRequestUpdateDto.getPayment().getPaymentMethod())
-                                                                  .caseReference(serviceRequestUpdateDto.getPayment().getCaseReference())
-                                                                  .accountNumber(serviceRequestUpdateDto.getPayment().getAccountNumber())
-                                                                  .build()).build())
-                                             .build(), CaseData.class);
+                 .id(Long.valueOf(serviceRequestUpdateDto.getCcdCaseNumber()))
+                  .paymentCallbackServiceRequestUpdate(CcdPaymentServiceRequestUpdate.builder()
+                      .serviceRequestReference(serviceRequestUpdateDto.getServiceRequestReference())
+                       .ccdCaseNumber(serviceRequestUpdateDto.getCcdCaseNumber())
+                       .serviceRequestAmount(serviceRequestUpdateDto.getServiceRequestAmount())
+                       .serviceRequestStatus(serviceRequestUpdateDto.getServiceRequestStatus())
+                       .callBackUpdateTimestamp(LocalDateTime.now())
+                        .payment(CcdPayment.builder().paymentAmount(
+                                  serviceRequestUpdateDto.getPayment().getPaymentAmount())
+                                  .paymentReference(serviceRequestUpdateDto.getPayment().getPaymentReference())
+                                  .paymentMethod(serviceRequestUpdateDto.getPayment().getPaymentMethod())
+                                  .caseReference(serviceRequestUpdateDto.getPayment().getCaseReference())
+                                     .accountNumber(serviceRequestUpdateDto.getPayment().getAccountNumber())
+                                      .build()).build()).build(), CaseData.class);
     }
 }
